updated database

+++ js/database-file-functions.js
@@ -1,24 +1,17 @@
-var fileButton = document.querySelector('#fileButton');
 var storageRef = firebase.storage();
+var fileButton = document.querySelector('#fileButton');
+var file_status = document.querySelector("#file_status");
 
-fileButton.addEventListener('change', function(e) {
-  var file_status = document.getElementById("file_status");
+//--------------------------------------------
+//-----------------UPLOAD---------------------
+//--------------------------------------------
 
-  //CHECK LOGIN
-  var user = firebase.auth().currentUser;
+fileButton.addEventListener('change', function(e) { uploadFile(e) });
 
-  if (!user) {
-    // User is not signed in.
-    file_status.innerHTML = "Būsena: KLAIDA - NEPRISIJUNGTA";
-    return;
-  }
-
-
-  //get files
+function uploadFile(e) {
   var file = e.target.files[0];
-  //create storage ref
-  var fileRef = storageRef.ref('uploads/' + file.name);
-  //upload
+  var fileRef = storageRef.ref("openData/" + file.name);
+
   var task = fileRef.put(file);
   task.on('state_changed',
     function progress(snapshot) {
@@ -30,34 +23,35 @@ fileButton.addEventListener('change', function(e) {
     },
     function complete() {
       file_status.innerHTML = "Būsena: FAILAS ĮKELTAS";
-
-      //put uploaded file data to the database
-      var date = moment().format('YYYY-MM-DD HH:mm:ss');
-      var docRef = firestore.doc("openData/" + date);
-      fileRef.getDownloadURL().then(function(url) {
-        docRef.set( {
-          name: file.name,
-          downloadURL: url,
-          date: date
-        }).then(function() {
-          file_status.innerHTML = "Būsena: ĮKELTA";
-        }).catch(function(error) {
-          file_status.innerHTML = "Būsena: KLAIDA - " + error.message;
-        });
-      });
+      updateDatabase(fileRef, file);
+      loadMessageList();
     }
   );
-});
-
-function delete_files(database_file_id, storage_file_name) {
-  firestore.doc("openData/" + database_file_id).delete();
-  var fileRef = storageRef.ref('uploads/' + storage_file_name);
-
-  fileRef.delete().then(function() {
-    load_files_button.click();
-  })
 }
 
-function download_files(storage_file_name) {
+function updateDatabase(fileRef, file) {
+  var date = moment().format('YYYY-MM-DD HH:mm:ss');
+  var docRef = firestore.doc("openData/" + date);
+  fileRef.getDownloadURL().then(function(url) {
+    docRef.set( {
+      name: file.name,
+      link: "sample",
+      downloadURL: url,
+      date: date
+    }).then(function() {
+      file_status.innerHTML = "Būsena: ĮKELTA";
+    }).catch(function(error) {
+      file_status.innerHTML = "Būsena: KLAIDA - " + error.message;
+    });
+  });
+}
 
+//--------------------------------------------
+//-----------------DELETE---------------------
+//--------------------------------------------
+
+function delete_files(database_file_id, storage_file_name) {
+  if (!firebase.auth().currentUser) return;
+  firestore.doc("openData/" + database_file_id).delete();
+  storageRef.ref('uploads/' + storage_file_name).delete().then(loadMessageList());
 }
