+++ js/database-message-functions.js
@@ -5,26 +5,21 @@
 const inputText = document.querySelector("#sendTextInput");
 const saveButton = document.querySelector("#sendTextButton");
 const linkText = document.querySelector('#sendLinkInput');
+const text_status = document.getElementById("text_status");
 
 saveButton.addEventListener("click", function () {
-  var text_status = document.getElementById("text_status");
-  //CHECK LOGIN
-  var user = firebase.auth().currentUser;
-
-  if (!user) {
-    // User is not signed in.
+  if (!firebase.auth().currentUser) {
     text_status.innerHTML = "Būsena: KLAIDA - NEPRISIJUNGTA";
     return;
   }
 
   text_status.innerHTML = "Būsena: SIUNČIAMA";
-  var date = moment().format('YYYY-MM-DD HH:mm:ss');
   var docRef = firestore.doc("openData/" + date);
   docRef.set( {
     name: inputText.value,
     link: linkText.value,
     downloadURL: "sample",
-    date: date
+    date: moment().format('YYYY-MM-DD HH:mm:ss')
   }).then(function() {
     text_status.innerHTML = "Būsena: NUSIŲSTA";
   }).catch(function(error) {
@@ -40,31 +35,56 @@ const load_messages_button = document.querySelector('#load_messages_button');
 const load_messages_list = document.querySelector('#load_messages_list');
 
 load_messages_button.addEventListener("click", function() {
+  loadMessageList();
+});
+
+function loadMessageList() {
   firestore.collection('openData').get().then(function(snap) {
     var idx = 1;
     load_messages_list.innerHTML = "";
     snap.forEach(function(doc) {
-      if(doc.data().downloadURL != "sample") {
-        var text = "<li><b>" + idx + ". " + doc.data().name + '</b><br>' + doc.data().date + '<br><u onclick="delete_files(' + "'" + doc.id + "','" + doc.data().name + "'" + ')">IŠTRINTI</u>, <a href="' + doc.data().downloadURL + '">SIŲSTIS</a>' + "</li>"
+      if(doc.data().name != "sample") {
+        var link = doc.data().link, name = doc.data().name, date = doc.data().date, url = doc.data().downloadURL, id = doc.id;
+
+        var text = "<li";
+        if(idx % 2 == 0)
+          text += ' class="lyginis"';
+        text += "><b>" + idx + ". ";
+
+        if(link != "sample")
+          text += '<a href="' + link + '">' + name + "</a>";
+        else text += name;
+
+        text += "</b><br>" + date + '<br><u onclick="';
+
+        if(url != "sample")
+          text += 'delete_files(' + "'" + doc.id + "','" + name + "'";
+        else
+          text += 'delete_messages(' + "'" + doc.id + "'";
+
+        text += ')">IŠTRINTI</u>';
+
+        if(url != "sample")
+          text += ', <a href="' + url + '">SIŲSTIS</a>';
+        text += "</li>"
+
         load_messages_list.innerHTML += text;
-      } else if(doc.data().name != "sample") {
-        var text = "<li><b>" + idx + ". ";
-        if(doc.data().link != "sample")
-          text += '<a href="' + doc.data().link + '">' + doc.data().name + "</a>";
-        else text += doc.data().name;
-        text += '</b><br>' + doc.data().date + '<br><u onclick="delete_messages(' + "'" + doc.id + "'" + ')">IŠTRINTI</u>' + "</li>"
-        load_messages_list.innerHTML += text;
+
+        idx++;
       }
-      idx++;
     });
   });
-});
+}
 
 //--------------------------------------------
 //-----------------DELETE---------------------
 //--------------------------------------------
 
 function delete_messages(document_id) {
+  if (!firebase.auth().currentUser) {
+    text_status.innerHTML = "Būsena: KLAIDA - NEPRISIJUNGTA";
+    return;
+  }
   firestore.doc("openData/" + document_id).delete();
-  load_messages_button.click();
+  loadMessageList();
 }
